pyproject.toml
[project]
name = "fastapi_fitness_tracker"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy[asyncio] (>=2.0.38,<3.0.0)",
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "celery[redis] (>=5.4.0,<6.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "unidecode (>=1.3.8,<2.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
